requirements.txt
pydantic==2.10.6
gradio
google-genai
python-dotenv
sqlalchemy
pandas
plotly
seaborn